+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## Unreleased
 
 ### Enhancements
+- New roles to deploy Remote Desktop Services
+- VM path now globally configurable
 
 ### Fixes
 - Fixed #1087. AL now supports WIM and ESD files in Windows ISOs.
 - Fixed #1115, the validator for AzDevOps/TFS now also validates SQLServer2019
+- Fixing issue with Get-LabInternetFile on Azure when no LabSources have been synced yet.
 
 ## 5.33.0 (2021-03-03)